# src/main/java/com/ngominhthuan/product_and_order_api/DTO/OrderRequest.java
package com.ngominhthuan.product_and_order_api.DTO;

import lombok.Data;
import java.util.List;

@Data
public class OrderRequest {
    private Long userId;
    private List<OrderItemRequest> items;

    @Data
    public static class OrderItemRequest {
        private Long productId;
        private Integer quantity;
    }
}