# src/main/java/com/ngominhthuan/product_and_order_api/DTO/UserResponse.java
package com.ngominhthuan.product_and_order_api.DTO;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserResponse {
    private Long id;
    private String username;
    private String email;
    private String role;
}