# src/main/resources/static/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Product and Order API",
    "description": "A REST API for managing products and orders with JWT authentication. This system provides secure endpoints for product catalog management, user authentication, and order processing with automatic transaction handling.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@lightvision.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "User registration and login management with JWT token generation"
    },
    {
      "name": "Product Management",
      "description": "CRUD operations for product catalog and inventory management"
    },
    {
      "name": "Order Management",
      "description": "Order placement and transaction processing with automatic rollback"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Creates a new user account with encrypted password using BCrypt hashing algorithm",
        "operationId": "register",
        "requestBody": {
          "description": "User registration details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "username": "johndoe",
                "password": "SecurePass123!",
                "email": "johndoe@example.com",
                "role": "USER"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "id": 1,
                  "username": "johndoe",
                  "email": "johndoe@example.com",
                  "role": "USER"
                }
              }
            }
          },
          "400": {
            "description": "Registration failed - Username or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "usernameExists": {
                    "summary": "Username already taken",
                    "value": {
                      "message": "Username is already taken"
                    }
                  },
                  "emailExists": {
                    "summary": "Email already taken",
                    "value": {
                      "message": "Email is already taken"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate user",
        "description": "Validates user credentials and returns a JWT token for subsequent authenticated requests",
        "operationId": "login",
        "requestBody": {
          "description": "User login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "username": "johndoe",
                "password": "SecurePass123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful - JWT token returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJqb2huZG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
              }
            }
          },
          "401": {
            "description": "Authentication failed - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Incorrect username or password"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Product Management"],
        "summary": "Retrieve all products",
        "description": "Returns a complete list of all products in the catalog with their current stock levels. Requires authentication with valid JWT token.",
        "operationId": "getAllProducts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved product list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "name": "Laptop Pro 15",
                    "price": 1299.99,
                    "stock": 50,
                    "description": "High-performance laptop with 15-inch display"
                  },
                  {
                    "id": 2,
                    "name": "Wireless Mouse",
                    "price": 29.99,
                    "stock": 200,
                    "description": "Ergonomic wireless mouse with 2.4GHz connectivity"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Product Management"],
        "summary": "Create a new product",
        "description": "Adds a new product to the catalog. Requires authentication with valid JWT token. (ADMIN role required)",
        "operationId": "createProduct",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Product details to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              },
              "example": {
                "name": "Mechanical Keyboard",
                "price": 89.99,
                "stock": 75,
                "description": "RGB mechanical keyboard with blue switches"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 3,
                  "name": "Mechanical Keyboard",
                  "price": 89.99,
                  "stock": 75,
                  "description": "RGB mechanical keyboard with blue switches"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Product Management"],
        "summary": "Get product by ID",
        "description": "Retrieves detailed information about a specific product using its unique identifier. Requires authentication with valid JWT token.",
        "operationId": "getProductById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique product identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Product found and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 1,
                  "name": "Laptop Pro 15",
                  "price": 1299.99,
                  "stock": 50,
                  "description": "High-performance laptop with 15-inch display"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found with the specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Product Management"],
        "summary": "Delete a product",
        "description": "Permanently removes a product from the catalog. Requires authentication with valid JWT token. (ADMIN role required)",
        "operationId": "deleteProduct",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique product identifier to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found with the specified ID"
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "tags": ["Order Management"],
        "summary": "Place a new order",
        "description": "Creates an order for a user with automatic inventory management. Features transaction rollback if any item is out of stock, ensuring data consistency. Requires authentication.",
        "operationId": "createOrder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Order details including user ID and list of items",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              },
              "example": {
                "userId": 1,
                "items": [
                  {
                    "productId": 1,
                    "quantity": 2
                  },
                  {
                    "productId": 2,
                    "quantity": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed successfully with inventory updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": 1,
                  "user": {
                    "id": 1,
                    "username": "johndoe",
                    "email": "johndoe@example.com",
                    "role": "USER"
                  },
                  "totalAmount": 2629.97,
                  "status": "PENDING",
                  "createdAt": "2025-01-15T10:30:00",
                  "orderItems": [
                    {
                      "id": 1,
                      "product": {
                        "id": 1,
                        "name": "Laptop Pro 15",
                        "price": 1299.99,
                        "stock": 48,
                        "description": "High-performance laptop"
                      },
                      "quantity": 2,
                      "price": 1299.99
                    },
                    {
                      "id": 2,
                      "product": {
                        "id": 2,
                        "name": "Wireless Mouse",
                        "price": 29.99,
                        "stock": 199,
                        "description": "Ergonomic wireless mouse"
                      },
                      "quantity": 1,
                      "price": 29.99
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Order creation failed - Insufficient stock or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "outOfStock": {
                    "summary": "Product out of stock",
                    "value": {
                      "message": "Product 'Laptop Pro 15' is out of stock"
                    }
                  },
                  "userNotFound": {
                    "summary": "User not found",
                    "value": {
                      "message": "User not found"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "email"],
        "properties": {
          "username": {
            "type": "string",
            "description": "Unique username for the account",
            "minLength": 3,
            "maxLength": 50,
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "description": "Password (will be encrypted with BCrypt)",
            "format": "password",
            "minLength": 6,
            "example": "SecurePass123!"
          },
          "email": {
            "type": "string",
            "description": "User's email address",
            "format": "email",
            "example": "johndoe@example.com"
          },
          "role": {
            "type": "string",
            "description": "User role in the system",
            "enum": ["USER", "ADMIN"],
            "default": "USER",
            "example": "USER"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "Username for authentication",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "format": "password",
            "example": "SecurePass123!"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique user identifier",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "johndoe@example.com"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "example": "USER"
          }
        }
      },
      "AuthResponse": {
        "type": "string",
        "description": "JWT authentication token (use in Authorization header as 'Bearer <token>')",
        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJqb2huZG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
      },
      "ProductRequest": {
        "type": "object",
        "required": ["name", "price", "stock"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "minLength": 1,
            "maxLength": 255,
            "example": "Laptop Pro 15"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price in USD",
            "minimum": 0,
            "example": 1299.99
          },
          "stock": {
            "type": "integer",
            "format": "int32",
            "description": "Available quantity in inventory",
            "minimum": 0,
            "example": 50
          },
          "description": {
            "type": "string",
            "description": "Detailed product description",
            "maxLength": 5000,
            "example": "High-performance laptop with 15-inch display, 16GB RAM, and 512GB SSD"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique product identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Laptop Pro 15"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Product price in USD",
            "example": 1299.99
          },
          "stock": {
            "type": "integer",
            "format": "int32",
            "description": "Available quantity in inventory",
            "example": 50
          },
          "description": {
            "type": "string",
            "description": "Detailed product description",
            "example": "High-performance laptop with 15-inch display"
          }
        }
      },
      "OrderItemRequest": {
        "type": "object",
        "required": ["productId", "quantity"],
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the product to order",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "description": "Quantity to order",
            "minimum": 1,
            "example": 2
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "required": ["userId", "items"],
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the user placing the order",
            "example": 1
          },
          "items": {
            "type": "array",
            "description": "List of products and quantities to order",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/OrderItemRequest"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique order item identifier",
            "example": 1
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "description": "Quantity ordered",
            "example": 2
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Price at the time of order",
            "example": 1299.99
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique order identifier",
            "example": 1
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "description": "Total order amount in USD",
            "example": 2629.97
          },
          "status": {
            "type": "string",
            "description": "Current order status",
            "enum": ["PENDING", "COMPLETED", "CANCELLED"],
            "example": "PENDING"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Order creation timestamp",
            "example": "2025-01-15T10:30:00"
          },
          "orderItems": {
            "type": "array",
            "description": "List of items in the order",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Username already exists!"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from the login endpoint. Format: Bearer <token>"
      }
    }
  }
}